Operações básicas para entidade Spent

+++ src/main/kotlin/com/budget/api/service/UserService.kt
@@ -1,11 +1,10 @@
 package com.budget.api.service
 
 import com.budget.api.message.request.UserRequest
-import com.budget.api.message.response.error.ErrorResponse
-import com.budget.api.message.response.success.UserResponse
 import com.budget.api.model.User
+import com.budget.api.repository.SpentRepository
 import com.budget.api.repository.UserRepository
-import com.budget.api.service.exception.UserException
+import com.budget.api.service.exception.BudgetException
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.stereotype.Service
 import java.util.*
@@ -16,6 +15,9 @@ import java.util.regex.Pattern
 class UserService {
     @Autowired
     lateinit var userRepository: UserRepository
+
+    @Autowired
+    lateinit var spentRepository: SpentRepository
 
     /*
     *  Método responsável por verificar se o cpf já existe
@@ -76,7 +78,7 @@ class UserService {
         val user = userRepository.findById(id)
 
         if (!user.isPresent) {
-            throw UserException(404, "Usuário não encontrado")
+            throw BudgetException(404, "Usuário não encontrado")
         }
 
         return user
@@ -88,7 +90,13 @@ class UserService {
     * @params id
     */
     fun deleteById(id: Long) {
-        return userRepository.deleteById(id)
+        val user = userRepository.findById(id)
+
+        if (!user.isPresent) {
+            throw BudgetException(404, "Usuário não encontrado")
+        }
+
+        userRepository.deleteById(id)
     }
 
     /*
@@ -116,7 +124,7 @@ class UserService {
     */
     fun validatePassword(user: User) {
         if (user.password?.length!! < 8) {
-            throw UserException(400, "Senha deve ser maior que 8 caracteres")
+            throw BudgetException(400, "Senha deve ser maior que 8 caracteres")
         }
     }
 
@@ -130,35 +138,35 @@ class UserService {
         var matcher: Matcher = pattern.matcher(user.email)
 
         if (user.name?.isEmpty()!!) {
-            throw UserException(400, "Nome obrigatório")
+            throw BudgetException(400, "Nome obrigatório")
         }
 
         if (user.name?.length!! < 3 || user.name?.length!! > 80) {
-            throw UserException(400, "Nome deve ter tamanho entre 3 e 80 caracteres")
+            throw BudgetException(400, "Nome deve ter tamanho entre 3 e 80 caracteres")
         }
 
         if (user.cpf?.isEmpty()!! || user.cpf?.length != 11) {
-            throw UserException(400, "CPF inválido")
+            throw BudgetException(400, "CPF inválido")
         }
 
         if (user.email?.isEmpty()!!) {
-            throw UserException(400, "Email obrigatório")
+            throw BudgetException(400, "Email obrigatório")
         }
 
         if (user.income == null) {
-            throw UserException(400, "Renda inválida")
+            throw BudgetException(400, "Renda inválida")
         }
 
         if (!matcher.matches()) {
-            throw UserException(400, "Email inválido")
+            throw BudgetException(400, "Email inválido")
         }
 
         if (userRepository.existsByCpf(user.cpf!!)) {
-            throw UserException(400, "CPF já cadastrado")
+            throw BudgetException(400, "CPF já cadastrado")
         }
 
         if (userRepository.existsByEmail(user.email!!)) {
-            throw UserException(400, "Email já cadastrado")
+            throw BudgetException(400, "Email já cadastrado")
         }
 
         validatePassword(user)
